test: RestaurantReviewStatServiceTest를 제거하고 RestaurantServiceTest로 통일
